CORE-8711 Move ManageCollaboratorsView interface to client package for consistency
