+++ de-lib/src/main/java/org/iplantc/de/collaborators/client/views/ManageCollaboratorsDialog.java
@@ -1,7 +1,8 @@
 package org.iplantc.de.collaborators.client.views;
 
 import org.iplantc.de.client.models.collaborators.Collaborator;
-import org.iplantc.de.collaborators.client.views.ManageCollaboratorsView.MODE;
+import org.iplantc.de.collaborators.client.ManageCollaboratorsView;
+import org.iplantc.de.collaborators.client.ManageCollaboratorsView.MODE;
 import org.iplantc.de.collaborators.shared.CollaboratorsModule;
 import org.iplantc.de.commons.client.views.dialogs.IPlantDialog;
