+++ de-lib/src/main/java/org/iplantc/de/theme/base/client/collaborators/Collaborators.gwt.xml
@@ -5,7 +5,7 @@
     <inherits name="org.iplantc.de.collaborators.Collaborators"/>
 
     <replace-with class="org.iplantc.de.theme.base.client.collaborators.ManageCollaboratorsViewDefaultAppearance">
-        <when-type-is class="org.iplantc.de.collaborators.client.views.ManageCollaboratorsView.Appearance" />
+        <when-type-is class="org.iplantc.de.collaborators.client.ManageCollaboratorsView.Appearance" />
     </replace-with>
 
     <replace-with class="org.iplantc.de.theme.base.client.collaborators.GroupViewDefaultAppearance">
