+++ de-lib/src/main/java/org/iplantc/de/theme/base/client/collaborators/ManageCollaboratorsViewDefaultAppearance.java
@@ -1,6 +1,6 @@
 package org.iplantc.de.theme.base.client.collaborators;
 
-import org.iplantc.de.collaborators.client.views.ManageCollaboratorsView;
+import org.iplantc.de.collaborators.client.ManageCollaboratorsView;
 import org.iplantc.de.resources.client.IplantResources;
 import org.iplantc.de.resources.client.messages.IplantContextualHelpStrings;
 import org.iplantc.de.resources.client.messages.IplantDisplayStrings;
